ratings slider, datenbank aktualisierung, und noch as

+++ BoOp.DBAccessor/Models/BasicRezensionenModel.cs
@@ -11,7 +11,7 @@
     {
         public int? Id { get; set; }
         public int BuchID { get; set; }
-        public int Sterne { get; set; }
+        public double Sterne { get; set; }
         public string Rezensionstext { get; set; }
         public int PersonID { get; set; }
     }
