+++ BoOp.UI.WPF/Common/RatingViewModel.cs
@@ -6,14 +6,14 @@ namespace BoOp.UI.WPF.Common
     //Dateiname: RatingViewModel.cs
     //Author : Manuel Janzen
     //Erstellt am : 27/10/2021
-    //Bearbeitet von : Manuel Janzen
-    //Beschreibung : View Model für personen
+    //Bearbeitet von : Manuel Janzen, Dominik von Michalkowsky
+    //Beschreibung : View Model für bewertungen
     ////////////////////////////////////////////////////////////////////////////////////////////////////////
     public class RatingViewModel : ViewModel
     {
-        public int Rating { get; set; }
+        public double Rating { get; set; }
 
-        public RatingViewModel(int rating)
+        public RatingViewModel(double rating)
         {
             Rating = rating;
         }
